+++ apps/web/src/app/(main)/(docs)/components/input/variants/clearable.tsx
@@ -1,0 +1,12 @@
+import { Input } from "@jamsr-ui/react";
+
+export const InputClearable = () => {
+  return (
+    <Input
+      label="Username"
+      placeholder="Enter your username"
+      type="text"
+      isClearable
+    />
+  );
+};