+++ apps/web/src/app/(main)/(docs)/components/tab/variants/customized.tsx
@@ -1,9 +1,19 @@
 import { Tab, Tabs } from "@jamsr-ui/react";
 import { TabContent } from "../shared";
 
-export const TabUsage = () => {
+export const TabCustomized = () => {
   return (
-    <Tabs defaultValue="photos">
+    <Tabs
+      classNames={{
+        base: "bg-content2 rounded",
+        cursor: "rounded-full !bg-content-4",
+        panel: "bg-content3",
+        tab: "border border-gray-900",
+        // tabContent:"bg-content4"
+        // tabList:"bg-content4"
+      }}
+      defaultValue="photos"
+    >
       <Tab value="photos" heading="Photos">
         <TabContent />
       </Tab>