+++ packages/components/tab/src/tabs.tsx
@@ -3,7 +3,7 @@ import { useId } from "react";
 import { TabContext } from "./tabs-context";
 import { useTabs, type UseTabsProps } from "./use-tabs";
 
-export type TabsProps<T extends string> = UseTabsProps<T>;
+export type TabsProps<T extends string = never> = UseTabsProps<T>;
 
 export const Tabs = <T extends string>(props: TabsProps<T>) => {
   const {