+++ packages/components/tab/src/types.d.ts
@@ -1,7 +1,8 @@
-import { type TabProps } from ".";
+import { type TabProps, type TabsProps } from ".";
 
 declare module "@jamsr-ui/styles" {
   export interface UIStylesType {
     tab?: Partial<TabProps>;
+    tabs?: Partial<TabsProps>;
   }
 }