+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/input
 
+## 0.0.88
+
+### Patch Changes
+
+- fix input
+
 ## 0.0.87
 
 ### Patch Changes